requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.27.0

# Firebase
firebase-admin==6.4.0

# Vector database
chromadb==0.4.22

# OpenAI
openai==1.12.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9

# Scheduling
APScheduler==3.10.4

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# HTTP requests (for health checks)
httpx==0.27.0

# JSON handling
orjson==3.9.15